Assignments/Quarter-4/LangChain-RAG-Average/pyproject.toml
[tool.poetry]
name = "ragmain"
version = "0.1.0"
description = ""
authors = ["Abrar-Kamoka <abrarabbas.ak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.7"
langchain = "^0.3.13"
langchain-community = "^0.3.13"
langchain-core = "^0.3.28"
huggingface-hub = "^0.27.0"
bs4 = "^0.0.2"
faiss-cpu = "^1.9.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
